+++ CHANGELOG.md
@@ -1,13 +1,25 @@
 # Changelog
 
-## [0.1.45] - TBD
+
+## [0.1.46] - TBD
 ### Added
-- (*In progress*) A universal remote cache is automatically available for retrieving responses that have been previously run by any user at the Expected Parrot server (i.e., if you run a question that you or anyone else has run before, you will retrieve that response at no cost to you). It is available for all jobs run remotely by default, and new responses for any remote jobs are automatically added to it. If you want to draw fresh responses you can use `run(fresh_cache=True)` or `run(cache=Cache()`. If you want to draw responses from a different cache you can use `run(cache=<my_cache>)` (insert your own cache object). If you draw a fresh response for a question that has already been run, the new response is also added to the universal remote cache with an index (`n=2`, etc.). Universal remote cache is not available for jobs run locally. See the [remote cache](https://docs.expectedparrot.com/en/latest/remote_caching.html) section for more details.
+- (*In progress*) A universal remote cache is available for retrieving responses to any questions that have been run at the Expected Parrot server. If you re-run a question that anyone has run before, you can retrieve that response at no cost to you. This cache is available for all jobs run remotely by default, and new responses are automatically added to it. If you want to draw fresh responses you can use `run(fresh=True)`. If you draw a fresh response for a question that has already been run, the new response is also added to the URC with an iteration index. The URC is not available for jobs run locally. See the [remote cache](https://docs.expectedparrot.com/en/latest/remote_caching.html) section for more details.
+
+
+## [0.1.45] - 2025-02-27
+### Added
+- `ScenarioList` method `from_dta()` creates a scenario list from a Stata file.
+
+- `Results` method `flatten()` will flatten a field of dictionaries into separate fields. It takes a list of the fields to flatten and a boolean indicator whether to preserve the original fields in the new `Results` object that is returned. [See example](https://docs.expectedparrot.com/en/latest/results.html#flattening-resuls).
+
+- `Results` method `report()` generates a report of selected columns in markdown by iterating through the rows, presented as observations. You can optionally pass headers, a divider and a limit on the number of observations to include. It can be useful if you want to display some sample part of larger results in a working notebook you are sharing. [See example](https://docs.expectedparrot.com/en/latest/results.html#generating-a-report).
+
+- `Survey` method `show_flow()` can now also be called on a `Jobs` object, and will show any scenarios and/or agent traits that that you have added to questions. [See examples](https://docs.expectedparrot.com/en/latest/docs/surveys.html#show-flow).
 
 
 ## [0.1.44] - 2025-02-14
 ### Added
-- Grok models are now available. If you have your own key, you can add it to your Keys page at your Coop account or add `XAI_API_KEY=<your_key_here>` to your `.env` file.
+- Xai models are now available. If you have your own key, you can add it to your Keys page at your Coop account or add `XAI_API_KEY=<your_key_here>` to your `.env` file.
 
 - `Survey` method `humanize()` will create a web-based version of your survey to share with humans. Responses are automatically added to a `Results` object that you can access at your account. *This feature is live but in development.*  
 
@@ -34,7 +46,7 @@
 
 ## [0.1.42] - 2025-01-24
 ### Added
-- DeepSeek models, e.g., `Model("deepseek-reasoner")`.
+- DeepSeek models are now available (e.g., try `Model("deepseek-reasoner")`). If you have your own key, you can add it to your Keys page at your Coop account or add `DEEPSEEK_API_KEY=<your_key_here>` to your `.env` file.
 
 - The name of the inference service is now included in the `Model` parameters and `Results` objects. This can be useful when the same model is provided by multiple services.